[#13543] - Refactoring
